undo: Move checkpoint restoration under data.undo
Undo references and worktree capture now live under data.undo. The code
that restores a checkpoint still sits at data.undo_restore.

Restoration reads saved Git objects and rewrites worktree files, index
entries, references, and repository state. Those operations serve undo
checkpoints rather than a general data service.

This commit continues the package migration by moving restoration under
data.undo and redirecting checkpoint code to that location.

The next commit will update restoration tests before snapshot creation is
separated from stack movement.
